kafka_ws_connector_dir/requirements.txt
asyncio
websockets
aiokafka

# not sure things below
authlib
starlette
itsdangerous